requirements.txt
torch==1.9.0
torchvision==0.10.0
streamlit==0.87.0
pandas==1.2.4
numpy==1.20.3
Pillow==8.2.0
matplotlib==3.4.2
pycocotools==2.0.2
opencv-python
fastapi
uvicorn
python-multipart